optimize internal data structures
convert internal data structures to u8

Since non-multibyte alphabets and therefore non-multibyte strings are
not supported we can operate on bytes, i.e. `u8`, instead of `char`.
This is a purely internal change that is not visible in the external
interface. Changing the data types also required some slight refactoring
which got rid of some unnecessary allocations.