optimize blocklist
The blocklist is read into a HashSet twice. The blocklist that's
actually used does not need to be a HashSet since it's only iterated
over. By converting it into a vector we can save some memory and also
sort it by word length. This allows skipping words longer than the id
when checking for blocked ids. Blocklist words are never resized so we
can drop the capacity field by converting them into boxed slices. Since
we treat words containing digits differently, we can mark words
containing a digit when reading the blocklist.